*: move pkg to internal
Follow the golang conventions for internal packages and move them to the
internal dir.
