+++ cmd/keeper/cmd/keeper.go
@@ -32,14 +32,14 @@ import (
 	"time"
 
 	"github.com/sorintlab/stolon/cmd"
-	"github.com/sorintlab/stolon/common"
-	"github.com/sorintlab/stolon/pkg/cluster"
-	"github.com/sorintlab/stolon/pkg/flagutil"
-	slog "github.com/sorintlab/stolon/pkg/log"
-	"github.com/sorintlab/stolon/pkg/postgresql"
-	pg "github.com/sorintlab/stolon/pkg/postgresql"
-	"github.com/sorintlab/stolon/pkg/store"
-	"github.com/sorintlab/stolon/pkg/util"
+	"github.com/sorintlab/stolon/internal/cluster"
+	"github.com/sorintlab/stolon/internal/common"
+	"github.com/sorintlab/stolon/internal/flagutil"
+	slog "github.com/sorintlab/stolon/internal/log"
+	"github.com/sorintlab/stolon/internal/postgresql"
+	pg "github.com/sorintlab/stolon/internal/postgresql"
+	"github.com/sorintlab/stolon/internal/store"
+	"github.com/sorintlab/stolon/internal/util"
 
 	"github.com/davecgh/go-spew/spew"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
