+++ tests/integration/pitr_test.go
@@ -25,9 +25,9 @@ import (
 	"time"
 
 	"github.com/satori/go.uuid"
-	"github.com/sorintlab/stolon/common"
-	"github.com/sorintlab/stolon/pkg/cluster"
-	"github.com/sorintlab/stolon/pkg/store"
+	"github.com/sorintlab/stolon/internal/cluster"
+	"github.com/sorintlab/stolon/internal/common"
+	"github.com/sorintlab/stolon/internal/store"
 )
 
 func TestPITR(t *testing.T) {
