+++ tests/integration/sentinel_test.go
@@ -26,9 +26,9 @@ import (
 	"time"
 
 	uuid "github.com/satori/go.uuid"
-	"github.com/sorintlab/stolon/common"
-	"github.com/sorintlab/stolon/pkg/cluster"
-	"github.com/sorintlab/stolon/pkg/store"
+	"github.com/sorintlab/stolon/internal/cluster"
+	"github.com/sorintlab/stolon/internal/common"
+	"github.com/sorintlab/stolon/internal/store"
 )
 
 func TestSentinelEnabledProxies(t *testing.T) {
